Move upcast helper function to its own file
